feat: color span run results by optimization direction; align scope controls
- Hoist the duplicated computePositiveOptimization helper out of
  EvaluatorOutputPreview and CodeEvaluatorTestSection into the shared
  evaluators utils module.
- Color the span run result chip with the annotation's optimization
  direction and threshold — green above the pivot, red below, neutral when
  no direction can be determined — using the same optimization-direction
  tokens as the evaluator output previews, instead of unconditional green.
- Pin the sampling number field and slider row to the M input height so the
  Target toggles, sampling control, and Preview window select read as one
  control strip.
